+++ CHANGELOG.md
@@ -30,6 +30,7 @@ All notable changes to this project will be documented in this file. The format 
 ### Fixed
 - Fix mesh-expansion with one layer `mesh.expand(n=1)`. This expands the dimension of the points-array.
 - Fix VTK-compatible cells in `CubeArbitraryOrderHexahedron`.
+- Fix Cauchy-stress evaluation of `SolidBody` and `SolidBodyNearlyIncompressible` on a 2d-Field (plane stress): Automatic fall-back to Kirchhoff-stress and print a warning.
 
 ### Removed
 - Remove the deprecated old-style argument `move` in `dof.biaxial()`.